Added Dialog Pop-ups to catch certain error scenarios for User.
Workout Library and Scheduler Screen now each provide a pop up to the USer to instruct them to create the necessary exercises/workouts necessary for them to function properly.

Play workout menu now only outputs the error message to the user if no workouts are present, not providing the play button.